removed clang-reported warnings

+++ src/computers/MultirunComputer.cpp
@@ -117,7 +117,7 @@ namespace turb {
 
         for (size_t i = 0; i < single_pts.size(); ++i) {
             std::vector<double> norms = integrator->get_norms(single_pts[i]);
-            for (int j = 0; j < norms.size(); ++j) {
+            for (int j = 0; (unsigned)j < norms.size(); ++j) {
                 if (fabs(norms[j]) > 300) continue;
             }
 
@@ -129,7 +129,7 @@ namespace turb {
 
             current_filename += output_number.str();
 
-            for (int j = 0; j < norms.size(); ++j) {
+            for (int j = 0; (unsigned)j < norms.size(); ++j) {
                 current_filename += "-" + std::to_string(norms[j]);
                 std::cerr << norms[j] << " ";
             }
